extract answer mapping logic to util to reduce duplication
